+++ App.js
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import cache from './app/utils/cache';
 import logger from './app/utils/logger';
 import StandingsPage from './app/pages/StandingsPage';
 import { LocaleProvider } from './app/contexts/LocaleContext';
