+++ app/components/StandingsCard/MatchForm/MatchForm.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { View } from 'react-native';
 
 import styles from './MatchForm.styles';
-import FormCircle from './FormCircle';
+import FormCircle from '../FormCircle/FormCircle';
 
 export default function MatchForm({ form = new Array(5).fill(null), numberOfMatches = 5 }) {
   for (let index = 0; form.length < numberOfMatches; index++) form.push(null);
