+++ app/components/StandingsCard/MatchForm/MatchForm.styles.js
@@ -1,5 +1,6 @@
 import { StyleSheet } from 'react-native';
-import colors from '../../styles/colors';
+
+import colors from '../../../styles/colors';
 
 export default StyleSheet.create({
   container: {
