+++ app/components/StandingsCard/StandingsCard.styles.js
@@ -1,11 +1,12 @@
 import { StyleSheet } from 'react-native';
 
+import colors from '../../styles/colors';
+
 export default StyleSheet.create({
   groupTitle: {
     flex: 1,
-    paddingTop: 5,
-    paddingLeft: 5,
     fontWeight: 'bold',
+    marginVertical: 5,
   },
   content: {
     flexDirection: 'row',
@@ -13,5 +14,10 @@ export default StyleSheet.create({
   cardStyle: {
     paddingTop: 0,
     paddingHorizontal: 0,
+  },
+  delimiter: {
+    elevation: 2,
+    borderColor: colors.lightGray,
+    borderRightWidth: 0.25,
   },
 });
