+++ app/components/StandingsCard/StandingsTableKey/StandingsTableKey.styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import theme from '../../styles/theme';
+import theme from '../../../styles/theme';
 
 export default StyleSheet.create({
   tableKey: {
