+++ app/components/StandingsCard/TeamStats/TeamStats.styles.js
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
 
-import theme from '../../styles/theme';
-import colors from '../../styles/colors';
+import theme from '../../../styles/theme';
+import colors from '../../../styles/colors';
 
 export default StyleSheet.create({
   container: {
